requirements.txt
pandas
openpyxl
sqlalchemy
schedule
matplotlib
python-dotenv

pytest
